rename websocket folder
