+++ src/components/index.ts
@@ -20,4 +20,4 @@ export * from './overflowableText';
 export * from './snackbarProvider';
 export * from './topBar';
 export * from './treeViewFinder';
-export * from './websocket';
+export * from './notifications';
